Django/requirements.txt
Django==4.2.1
djangorestframework==3.14.0
django-filter==23.2